move and use range
